Backports for 3.1.1 (#13103)
* Fix tiny error in doc code-block (#12973)

* Add a ref for LegendLocation (#12939)

* Temporarily disable docker_from_wheel test in CI (#12963)

* Update In-kind sponsors list (#13054)

* Update In-kind sponsors list

* Update README.md

* Update first_steps_1.rst (#13084)

* Update first_steps_1.rst

* Update docs/bokeh/source/docs/first_steps/first_steps_1.rst


* Update first_steps_1.rst

Made the changes suggested!

* Update docs/bokeh/source/docs/first_steps/first_steps_1.rst


* Update first_steps_1.rst

Follow up edits

---------


* Allow `static/` symlinks in `BOKEH_ACTION=install pip install -e .` (#13065)

* Disable user select globally when dragging slider handles (#13082)

* Add tests for custom arrayable functions (#12892)

* Add tests for custom arrayable functions

* add more tests for arrayable

* fix ESLint error

* fix another ESLint error

* List installed software and versions in CI (#12899)

* List installed software in CI

* Don't require npm@8, install most recent one

* Allow to correctly sort international strings in DataTable (#12900)

* Don't assign model's name to HTMLElement.name (#12901)

* Use CSS-safe IDs in generated HTML (#12902)

* Use CSS-safe IDs in generated HTML

* Generate DOM elements with unique IDs in load_notebook()

* Don't double stroke in diagonal hatch patterns (#12903)

* Add bokehjs example for grid band patterns

* Don't double stroke in diagonal hatch patterns

* Update visual baselines

* updated commands with correct links to executable (#12910)

changed examples/app/slider.py to examples/server/app/sliders.py 

transform_markers.py is updated here but shows as marker_map here
https://docs.bokeh.org/en/latest/docs/dev_guide/setup.html

* Copy files if symlinking fails in 'pack' build task (#12928)

* Document lack of serialization support for decimal.Decimal under Serializer class (#12935)

* Add message in Serializer docstring addressing lack of support for decimal.Decimal.

* Remove trailing whitespace

* Remove trailing whitespace

* Delete test file

* Upload latest docs more cleanly (#12955)

* Upload latest docs more cleanly

* Missing return type

* Check for git tags before build (#12962)

* Check for git tags during build

* Update setup.py


---------


* Added troubleshooting for issue #13001 (#13002)

* Added troubleshooting for issue #13001

* update

* updated ReadMe file of examples folder (#13066)

* updated ReadMe file of examples folder

* Update examples/README.md


* Update examples/README.md


* Update examples/README.md


* removed intergration section from readme

* added linebreaks

* fixed widespace issues

* trimmed trailing widespace

---------


* Fix race conditions related to layout resize and updates (#13100)

* Fix a race condition between plot and canvas resize

* Ignore child model/view state inconsistencies for now

* Notify ready after render and layout is done

* Add a regression test for issue #13104

* manual eslint patch

* Make autocompleteinput.py example work with Pandas 2.0 (#13057)

* fix baseline

* Added metadata to line_on_off.py (#13003)

* Added metadata to line_on_off.py

* update

* update

* update

* update

* Tile source metadata (#13026)

* added accessible color palette

* Delete latex_schrodinger.png

* Delete latex_schrodinger@2x.png

* Add new thumbnails for updated palette

* Add metadata to text.py

* Revert changes from ths branch

* Update tile_source.py

* fix consistency & linter issues

* fix broken docs refs (#13086)

* fix broken docs refs

* pin sphinx to 5.3

* sphinx pin for 3.8

* Use libmamba solver in CI (#13061)

* Use libmamba solver in CI

* Disable "Post prepare environment" on Windows

* Correctly resolve custom host and port with server resources (#13041)

* Correctly resolve custom host and port with server resources

* Use pathlib.Path instead of a comment

* Fix argument types in _bundle_extensions()

* Remove support for BaseResources

* Add a release note

* Add support for Resources.clone()

* Include settings.dev in Resources.dev

Allows to specify:

BOKEH_DEV=true BOKEH_RESOURCES=server python ...

instead of:

BOKEH_DEV=true BOKEH_RESOURCES=server-dev python ...

* Fix `MultiRootStaticHandler` after PR #13041 (#13070)

* Fix MultiRootStaticHandler after PR #13041

* Work around unawaited async function calls

* Fix bokeh.sphinxext.util after PR #13041 (#13062)

* Increase the number of iterations in CSS safe ID generator (#13035)

* Add dropdowns to contributor docs troubleshooting (#13009)

* Simpler WebGL line shaders (#13039)

* Simplify webgl line shaders

* Update baseline images

* Add new image tests

* Add Path.is_relative_to() polyfill for Python 3.8

* [BUG] Fix broken link to tools guide (#13109)

* [BUG] Fix broken link to tools guide

#13108

* [BUG] Fix broken link to tools guide

#13108

---------

+++ setup.py
@@ -83,16 +83,18 @@ def install_js(packages: list[str]) -> None:
     if missing:
         die(BOKEHJS_INSTALL_FAIL.format(missing=", ".join(missing)))
 
-    if PKG_JS.exists():
-        rmtree(PKG_JS)
-    copytree(BUILD_JS, PKG_JS)
+    if not PKG_JS.is_symlink():
+        if PKG_JS.exists():
+            rmtree(PKG_JS)
+        copytree(BUILD_JS, PKG_JS)
 
-    if PKG_TSLIB.exists():
-        rmtree(PKG_TSLIB)
-    if BUILD_TSLIB.exists():
-        PKG_TSLIB.mkdir()
-        for lib_file in BUILD_TSLIB.glob("lib.*.d.ts"):
-            copy(lib_file, PKG_TSLIB)
+    if not PKG_TSLIB.is_symlink():
+        if PKG_TSLIB.exists():
+            rmtree(PKG_TSLIB)
+        if BUILD_TSLIB.exists():
+            PKG_TSLIB.mkdir()
+            for lib_file in BUILD_TSLIB.glob("lib.*.d.ts"):
+                copy(lib_file, PKG_TSLIB)
 
     new = set(
         ".".join([*Path(parent).relative_to(SRC_ROOT).parts, d])
@@ -118,10 +120,23 @@ def build_or_install_bokehjs(packages: list[str]) -> None:
         raise ValueError(f"Unrecognized action {action!r}")
     print(f"Used {bright(yellow(kind))} BokehJS from {loc}\n")
 
+def check_tags() -> None:
+    try:
+        tags = subprocess.check_output(("git", "tag", "-l")).split()
+        if len(tags) < 5: # arbitrary "too-low" cutoff, there will always be more than 5 tags, if there are any
+            die(bright(red(MISSING_TAGS)))
+    except Exception:
+        print(bright(yellow("!!! Could not check repo tags. Please ensure full tag history")))
+
 def die(x: str) -> NoReturn:
     print(f"{x}\n")
     sys.exit(1)
 
+MISSING_TAGS = """
+!!! This repository is missing git tags! Full tag history is required to build Bokeh.
+!!!
+!!! See https://docs.bokeh.org/en/latest/docs/dev_guide/setup.html#troubleshooting
+"""
 SUCCESS = f"{bright(green('Success!'))}\n"
 FAILED = f"{bright(red('Failed.'))}\n"
 BUILD_SUCCESS_MSG =f"{SUCCESS}\nBuild output:\n\n{{msg}}"
@@ -141,11 +156,13 @@ BUILD_FAIL_MSG = f"""{FAILED}\nERROR: 'node make build' returned the following
 
 class Build(build):  # type: ignore
     def run(self) -> None:
+        check_tags()
         build_or_install_bokehjs(self.distribution.packages)
         super().run()
 
 class EditableWheel(editable_wheel):  # type: ignore
     def run(self) -> None:
+        check_tags()
         build_or_install_bokehjs(self.distribution.packages)
         super().run()
